+++ config/webpack/loaders/eslint.js
@@ -5,7 +5,7 @@ module.exports = {
   use: {
     loader: 'eslint-loader',
     options: {
-      fix: true
+      fix: false
     }
   }
 };