+++ hbw/app/javascript/packs/hbw/components/buttons.js.jsx
@@ -7,13 +7,12 @@ modulejs.define(
   'HBWButtons',
   ['React', 'HBWButton', 'HBWError'],
   (React, Button, Error) => {
-    const Buttons = React.createClass({
-      displayName: 'HBWButtons',
+    class HBWButtons extends React.Component {
+      constructor (props, context) {
+        super(props, context);
+        props.setGuid();
 
-      getInitialState () {
-        this.props.setGuid();
-
-        return {
+        this.state = {
           buttons:       [],
           subscription:  this.createSubscription(),
           pollInterval:  5000,
@@ -27,9 +26,9 @@ modulejs.define(
           bpRunning:     false,
           fileUploading: false
         };
-      },
+      }
 
-      createSubscription () {
+      createSubscription = () => {
         const subscription = this.props.env.connection.subscribe({
           client: this.props.getComponentId(),
           path:   '/buttons',
@@ -54,37 +53,33 @@ modulejs.define(
             syncError: null,
             bpRunning: data.bp_running
           }));
-      },
+      };
 
-      submitButton (businessProcessCode) {
-        return this.props.env.connection.request({
-          url:    this.buttonsURL(),
-          method: 'POST',
+      submitButton = businessProcessCode => this.props.env.connection.request({
+        url:    this.buttonsURL(),
+        method: 'POST',
 
-          data: {
-            entity_code:       this.props.entityCode,
-            entity_type:       this.props.entityTypeCode,
-            entity_class:      this.props.entityClassCode,
-            bp_code:           businessProcessCode,
-            initial_variables: this.props.env.initialVariables
-          }
-        });
-      },
+        data: {
+          entity_code:       this.props.entityCode,
+          entity_type:       this.props.entityTypeCode,
+          entity_class:      this.props.entityClassCode,
+          bp_code:           businessProcessCode,
+          initial_variables: this.props.env.initialVariables
+        }
+      });
 
-      buttonsURL () {
-        return `${this.props.env.connection.serverURL}/buttons`;
-      },
+      buttonsURL = () => `${this.props.env.connection.serverURL}/buttons`;
 
       componentDidMount () {
         this.state.subscription.start(this.state.pollInterval);
         this.props.bind('hbw:button-activated', this.onButtonActivation);
         this.props.bind('hbw:file-upload-started', () => this.setState({ fileUploading: true }));
         this.props.bind('hbw:file-upload-finished', () => this.setState({ fileUploading: false }));
-      },
+      }
 
       componentWillUnmount () {
         this.state.subscription.close();
-      },
+      }
 
       render () {
         if (this.props.env.userExist) {
@@ -113,9 +108,9 @@ modulejs.define(
         } else {
           return <div></div>;
         }
-      },
+      }
 
-      onButtonActivation (button) {
+      onButtonActivation = (button) => {
         console.log(`Clicked button[${button.title}], submitting`);
         this.setState({
           syncing:    true,
@@ -135,13 +130,13 @@ modulejs.define(
             errorHeader: this.props.env.translator('errors.cannot_start_process')
           }))
           .always(() => this.setState({ syncing: false }));
-      },
+      };
 
-      triggerBPStart (button) {
+      triggerBPStart = (button) => {
         this.props.trigger('hbw:process-started', button);
-      }
-    });
+      };
+    }
 
-    return withCallbacks(Buttons);
+    return withCallbacks(HBWButtons);
   }
 );