+++ hbw/app/javascript/packs/hbw/components/helpers/conditions.js
@@ -1,16 +1,11 @@
 /* eslint-disable no-eval */
 
-import React from 'react';
-import { getDisplayName } from './utils';
+import React, { Component } from 'react';
 
-export default WrappedComponent => React.createClass({
-  displayName: `WithConditions(${getDisplayName(WrappedComponent)})`,
+export default WrappedComponent => class WithConditions extends Component {
+  variables = () => this.props.params.variables || this.props.params.task.definition.variables;
 
-  variables () {
-    return this.props.params.variables || this.props.params.task.definition.variables;
-  },
-
-  variableByName (name) {
+  variableByName = (name) => {
     const variable = this.variables().find(v => v.name === name);
 
     if (variable) {
@@ -18,17 +13,13 @@ export default WrappedComponent => React.createClass({
     }
 
     return null;
-  },
+  };
 
-  deleteIf () {
-    return this.evaluateConditions(this.getDeleteIfConditions());
-  },
+  deleteIf = () => this.evaluateConditions(this.getDeleteIfConditions());
 
-  disableIf () {
-    return this.evaluateConditions(this.getDisableIfConditions());
-  },
+  disableIf = () => this.evaluateConditions(this.getDisableIfConditions());
 
-  evaluateConditions (conditions) {
+  evaluateConditions = (conditions) => {
     if ((Array.isArray(conditions) && (conditions.length === 0)) || Object.keys(conditions).length === 0) {
       return false;
     }
@@ -44,46 +35,36 @@ export default WrappedComponent => React.createClass({
     }
 
     return this.every(result);
-  },
+  };
 
-  getDeleteIfConditions () {
-    return this.props.params.delete_if || [];
-  },
+  getDeleteIfConditions = () => this.props.params.delete_if || [];
 
-  getDisableIfConditions () {
-    return this.props.params.disable_if || [];
-  },
+  getDisableIfConditions = () => this.props.params.disable_if || [];
 
-  conditionType (condition) {
+  conditionType = (condition) => {
     if (condition.constructor === Array) {
       return 'or';
     } else {
       return 'and';
     }
-  },
+  };
 
-  evaluateAndCondition (data) {
-    return eval(data.condition.replace('$var', this.variableByName(data.variable)));
-  },
+  evaluateAndCondition = data => eval(data.condition.replace('$var', this.variableByName(data.variable)));
 
-  evaluateOrCondition (data) {
+  evaluateOrCondition = (data) => {
     const result = [...data].map(innerCondition => this.evaluateAndCondition(innerCondition));
 
     return this.every(result);
-  },
+  };
 
   componentWillMount () {
     this.hidden = this.deleteIf();
     this.disabled = this.disableIf();
-  },
+  }
 
-  every (results) {
-    return results.every(el => el !== false);
-  },
+  every = results => results.every(el => el !== false);
 
-  some (results) {
-    return results.some(el => el === true);
-  },
+  some = results => results.some(el => el === true);
 
   render () {
     return <WrappedComponent hidden={this.hidden}
@@ -91,4 +72,4 @@ export default WrappedComponent => React.createClass({
                              {...this.state}
                              {...this.props} />;
   }
-});
+};