+++ hbw/app/javascript/packs/hbw/components/menu_button.js.jsx
@@ -1,22 +1,18 @@
 import { withCallbacks, withTasks, compose } from './helpers';
 
 modulejs.define('HBWMenuButton', ['React'], (React) => {
-  const MenuButton = React.createClass({
-    displayName: 'HBWMenuButton',
+  class HBWMenuButton extends React.Component {
+    state = { tasksNumber: 0 };
 
-    getInitialState () {
-      return { tasksNumber: 0 };
-    },
-
-    toggleVisibility () {
+    toggleVisibility = () => {
       this.setState({ visible: !this.state.visible });
-    },
+    };
 
     componentDidMount () {
       this.props.bind('hbw:hide-widget', this.toggleVisibility);
       this.props.subscription
         .progress(data => this.setState({ tasksNumber: data.tasks.length }));
-    },
+    }
 
     render () {
       const counter = this.state.tasksNumber ? <span className="counter">{this.state.tasksNumber}</span> : '';
@@ -26,15 +22,15 @@ modulejs.define('HBWMenuButton', ['React'], (React) => {
                 className="hbw-menu-button fas fa-reorder">
         {counter}
       </a>;
-    },
+    }
 
-    toggleMenu (e) {
+    toggleMenu = (e) => {
       e.preventDefault();
 
       this.rootNode.blur();
       this.props.trigger('hbw:toggle-tasks-menu');
-    }
-  });
+    };
+  }
 
-  return compose(withTasks, withCallbacks)(MenuButton);
+  return compose(withTasks, withCallbacks)(HBWMenuButton);
 });